refactor: replace AppLightColors with context theme colors across auth , discount and messages features
