+++ lib/features/authorization/presentation/views/personal_info.dart
@@ -1,7 +1,7 @@
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/features/authorization/presentation/widget/sign_up_have_an_account_section.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
-import 'package:aggar/core/themes/app_light_colors.dart';
 import 'package:aggar/core/utils/app_styles.dart';
 import 'package:aggar/core/widgets/custom_elevated_button.dart';
 import 'package:aggar/features/authorization/presentation/widget/custom_text_from_felid.dart';
@@ -101,7 +101,7 @@ class _PersonalInfoPageState extends State<PersonalInfoPage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: AppLightColors.myWhite100_1,
+      backgroundColor: context.theme.white100_1,
       body: SafeArea(
         child: SingleChildScrollView(
           child: Padding(
