+++ lib/features/authorization/presentation/widget/card_type.dart
@@ -1,4 +1,4 @@
-import 'package:aggar/core/themes/app_light_colors.dart';
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/core/utils/app_styles.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
@@ -27,14 +27,10 @@ class CardType extends StatelessWidget {
         decoration: BoxDecoration(
           border: Border.all(
             width: 2,
-            color: isSelected
-                ? AppLightColors.myBlue100_2
-                : AppLightColors.myBlack25,
+            color: isSelected ? context.theme.blue100_2 : context.theme.black25,
           ),
           borderRadius: BorderRadius.circular(10),
-          color: isSelected
-              ? AppLightColors.myBlue10_2
-              : AppLightColors.myWhite100_1,
+          color: isSelected ? context.theme.blue10_2 : context.theme.white100_1,
         ),
         padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 16),
         margin: const EdgeInsets.all(5),
@@ -44,25 +40,24 @@ class CardType extends StatelessWidget {
             Icon(
               icon ?? Icons.person,
               size: 50,
-              color: isSelected
-                  ? AppLightColors.myBlue100_2
-                  : AppLightColors.myBlack25,
+              color:
+                  isSelected ? context.theme.blue100_2 : context.theme.black25,
             ),
             const Gap(6),
             Text(
               title ?? 'User',
               style: AppStyles.regular20(context).copyWith(
                 color: isSelected
-                    ? AppLightColors.myBlue100_2
-                    : AppLightColors.myBlack25,
+                    ? context.theme.blue100_2
+                    : context.theme.black25,
               ),
             ),
             Text(
               subtitle ?? 'Can use cars & buy for them',
               style: AppStyles.regular12(context).copyWith(
                 color: isSelected
-                    ? AppLightColors.myBlue100_2
-                    : AppLightColors.myBlack25,
+                    ? context.theme.blue100_2
+                    : context.theme.black25,
               ),
               textAlign: TextAlign.center,
             ),
