+++ lib/features/authorization/presentation/widget/pick_image_icon_with_title_and_subtitle.dart
@@ -1,8 +1,8 @@
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:flutter/material.dart';
 import 'dart:io';
 import 'package:image_picker/image_picker.dart';
 import 'package:aggar/core/utils/app_assets.dart';
-import 'package:aggar/core/themes/app_light_colors.dart';
 import 'package:aggar/core/utils/app_styles.dart';
 import 'package:aggar/core/widgets/custom_icon.dart';
 
@@ -91,7 +91,7 @@ class _PickImageIconWithTitleAndSubtitleState
           child: Column(
             children: [
               CircleAvatar(
-                backgroundColor: AppLightColors.myBlue100_2,
+                backgroundColor: context.theme.blue100_2,
                 radius: 90,
                 child: _localImagePath != null
                     ? ClipRRect(
@@ -116,13 +116,13 @@ class _PickImageIconWithTitleAndSubtitleState
                     ? "Change profile photo"
                     : "Profile photo",
                 style: AppStyles.semiBold24(context).copyWith(
-                  color: AppLightColors.myBlue100_1,
+                  color: context.theme.blue100_1,
                 ),
               ),
               Text(
                 "Please make sure that the photo you \n    upload will not be modified later",
                 style: AppStyles.regular12(context).copyWith(
-                  color: AppLightColors.myBlack50,
+                  color: context.theme.black50,
                 ),
               ),
             ],
