+++ lib/features/authorization/presentation/widget/sign_in_face_book_and_google_buttons.dart
@@ -1,5 +1,5 @@
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/core/utils/app_assets.dart';
-import 'package:aggar/core/themes/app_light_colors.dart';
 import 'package:aggar/core/widgets/custom_icon.dart';
 import 'package:aggar/features/authorization/presentation/widget/sign_in_social_button.dart';
 import 'package:flutter/material.dart';
@@ -16,7 +16,7 @@ class SignInFaceBookAndGoogleButtons extends StatelessWidget {
       mainAxisAlignment: MainAxisAlignment.center,
       children: [
         SignInSocialButton(
-          textColor: AppLightColors.myGray100_1,
+          textColor: context.theme.gray100_1,
           icon: const CustomIcon(
             hight: 21,
             width: 21,
@@ -28,14 +28,14 @@ class SignInFaceBookAndGoogleButtons extends StatelessWidget {
         ),
         const Gap(20),
         SignInSocialButton(
-          textColor: AppLightColors.myBlack100,
+          textColor: context.theme.black100,
           icon: const CustomIcon(
             hight: 21,
             width: 21,
             flag: false,
             imageIcon: AppAssets.assetsIconsGoogleIcon,
           ),
-          color: AppLightColors.myGray100_1,
+          color: context.theme.gray100_1,
           text: 'Google',
         ),
       ],
