+++ lib/features/authorization/presentation/widget/sign_in_forget_password_button.dart
@@ -1,4 +1,4 @@
-import 'package:aggar/core/themes/app_light_colors.dart';
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/core/utils/app_styles.dart';
 import 'package:flutter/material.dart';
 
@@ -23,7 +23,7 @@ class SignInForgetPasswordButton extends StatelessWidget {
           child: Text(
             "Forget password?",
             style: AppStyles.medium18(context).copyWith(
-              color: AppLightColors.myBlue100_1,
+              color: context.theme.blue100_1,
             ),
           ),
         ),
