+++ lib/features/authorization/presentation/widget/sign_up_have_an_account_section.dart
@@ -1,4 +1,4 @@
-import 'package:aggar/core/themes/app_light_colors.dart';
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/core/utils/app_styles.dart';
 import 'package:flutter/material.dart';
 
@@ -27,7 +27,7 @@ class SignUpHaveAnAccountSection extends StatelessWidget {
           child: Text(
             "Login now",
             style: AppStyles.medium18(context).copyWith(
-              color: AppLightColors.myBlue100_1,
+              color: context.theme.blue100_1,
             ),
           ),
         ),
