+++ lib/features/authorization/presentation/widget/terms_check.dart
@@ -1,5 +1,5 @@
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:flutter/material.dart';
-import 'package:aggar/core/themes/app_light_colors.dart';
 import 'package:aggar/core/utils/app_styles.dart';
 
 class TermsCheck extends StatelessWidget {
@@ -19,7 +19,7 @@ class TermsCheck extends StatelessWidget {
         Checkbox(
           value: isChecked,
           onChanged: (value) => onChanged(value ?? false),
-          activeColor: AppLightColors.myBlue100_2,
+          activeColor: context.theme.blue100_2,
         ),
         GestureDetector(
           onTap: () {
