+++ lib/features/discount/presentation/widgets/add_discount_form.dart
@@ -1,4 +1,4 @@
-import 'package:aggar/core/themes/app_light_colors.dart';
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/features/discount/presentation/widgets/add_discount_button_list.dart';
 import 'package:aggar/features/discount/presentation/widgets/note_discount_section.dart';
 import 'package:aggar/features/discount/presentation/widgets/number_of_days_and_persentage_row.dart';
@@ -15,10 +15,10 @@ class AddDiscountForm extends StatelessWidget {
         Container(
           padding: const EdgeInsets.all(16),
           decoration: BoxDecoration(
-            color: AppLightColors.myWhite100_1,
+            color: context.theme.white100_1,
             boxShadow: [
               BoxShadow(
-                color: AppLightColors.myBlack25,
+                color: context.theme.black25,
                 blurRadius: 4,
                 offset: const Offset(0, 0),
               ),
