+++ lib/features/discount/presentation/widgets/custom_suffix_widget.dart
@@ -1,4 +1,4 @@
-import 'package:aggar/core/themes/app_light_colors.dart';
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/core/utils/app_styles.dart';
 import 'package:flutter/material.dart';
 
@@ -18,14 +18,14 @@ class CustomSuffixWidget extends StatelessWidget {
         border: Border(
           left: BorderSide(
             width: 1,
-            color: AppLightColors.myBlack50,
+            color: context.theme.black50,
           ),
         ),
       ),
       child: Text(
         suffixText,
         style: AppStyles.medium15(context).copyWith(
-          color: AppLightColors.myBlack50,
+          color: context.theme.black50,
         ),
       ),
     );
