+++ lib/features/discount/presentation/widgets/note_discount_section.dart
@@ -1,4 +1,4 @@
-import 'package:aggar/core/themes/app_light_colors.dart';
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/core/utils/app_styles.dart';
 import 'package:flutter/material.dart';
 
@@ -15,7 +15,7 @@ class NoteDiscountSection extends StatelessWidget {
         Text(
           "Note:",
           style: AppStyles.medium14(context).copyWith(
-            color: AppLightColors.myBlack25,
+            color: context.theme.black25,
           ),
         ),
         Text(
@@ -23,7 +23,7 @@ class NoteDiscountSection extends StatelessWidget {
           maxLines: 2,
           'Number of days that apply this offer and a percentage must be able to divided by 5',
           style: AppStyles.medium12(context).copyWith(
-            color: AppLightColors.myBlack25,
+            color: context.theme.black25,
           ),
         ),
       ],
