+++ lib/features/messages/views/messages_status/presentation/views/no_messages_view.dart
@@ -1,5 +1,5 @@
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/core/utils/app_assets.dart';
-import 'package:aggar/core/themes/app_light_colors.dart';
 import 'package:aggar/core/utils/app_styles.dart';
 import 'package:flutter/material.dart';
 
@@ -24,13 +24,13 @@ class NoMessagesView extends StatelessWidget {
             Text(
               "no chats yet!",
               style: AppStyles.medium20(context).copyWith(
-                color: AppLightColors.myBlue100_1,
+                color: context.theme.blue100_1,
               ),
             ),
             Text(
               " start chating",
               style: AppStyles.medium20(context).copyWith(
-                color: AppLightColors.myBlue100_2,
+                color: context.theme.blue100_2,
               ),
             ),
           ],
