+++ lib/features/messages/views/messages_status/presentation/widgets/widgets/avatar_chat_view.dart
@@ -1,4 +1,4 @@
-import 'package:aggar/core/themes/app_light_colors.dart';
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:flutter/material.dart';
 
 class AvatarChatView extends StatelessWidget {
@@ -15,7 +15,7 @@ class AvatarChatView extends StatelessWidget {
         boxShadow: [
           BoxShadow(
             offset: const Offset(0, 4),
-            color: AppLightColors.myBlack25,
+            color: context.theme.black25,
             spreadRadius: 0,
             blurRadius: 4,
           )
