+++ lib/features/messages/views/personal_chat/presentation/widgets/send_messages_with_attach_section.dart
@@ -1,4 +1,4 @@
-import 'package:aggar/core/themes/app_light_colors.dart';
+import 'package:aggar/core/extensions/context_colors_extension.dart';
 import 'package:aggar/features/messages/views/personal_chat/presentation/widgets/type_message_text_field_with_send_button.dart';
 import 'package:flutter/material.dart';
 
@@ -10,7 +10,7 @@ class SendMessagesWithAttachSection extends StatelessWidget {
     return Container(
       height: MediaQuery.sizeOf(context).height * 0.11,
       width: double.infinity,
-      color: AppLightColors.myWhite100_1,
+      color: context.theme.white100_1,
       child: const TypeMessageTextFieldWithSendButton(),
     );
   }
